+++ stac_fastapi/tests/conftest.py
@@ -38,8 +38,7 @@ else:
         create_index_templates,
     )
 
-from stac_fastapi.extensions.core import (  # FieldsExtension,
-    ContextExtension,
+from stac_fastapi.extensions.core import (
     FieldsExtension,
     FilterExtension,
     SortExtension,
@@ -193,7 +192,6 @@ async def app():
             ),
             settings=settings,
         ),
-        ContextExtension(),
         SortExtension(),
         FieldsExtension(),
         QueryExtension(),
@@ -236,7 +234,6 @@ async def app_basic_auth():
             ),
             settings=settings,
         ),
-        ContextExtension(),
         SortExtension(),
         FieldsExtension(),
         QueryExtension(),